fix db heads?
